separate common from load tests properly
