+++ .github/workflows/test_local_sources.yml
@@ -1,5 +1,4 @@
-# Tests destinations that can run without credentials.
-# i.e. local postgres, duckdb, filesystem (with local fs/memory bucket)
+# Tests sources against a couple of local destinations
 
 name: src | rest_api, sql_database, filesystem
 
@@ -22,6 +21,10 @@ env:
 
   ACTIVE_DESTINATIONS: "[\"duckdb\", \"postgres\", \"filesystem\"]"
   ALL_FILESYSTEM_DRIVERS: "[\"memory\", \"file\"]"
+
+  # we need the secrets to inject the github token for the rest_api template tests
+  # we should not use it for anything else here 
+  DLT_SECRETS_TOML: ${{ secrets.DLT_SECRETS_TOML }}
 
 jobs:
   get_docs_changes:
@@ -82,22 +85,20 @@ jobs:
           path: .venv
           key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}-local-sources
 
+      - name: create secrets.toml
+        run: pwd && echo "$DLT_SECRETS_TOML" > tests/.dlt/secrets.toml
+
         # TODO: which deps should we enable? 
       - name: Install dependencies
         run: poetry install --no-interaction -E postgres -E duckdb -E parquet -E filesystem -E cli -E sql_database --with sentry-sdk --with pipeline 
-        
-        # we need sqlalchemy 2 for the sql_database tests
+
+        # we need sqlalchemy 2 for the sql_database tests, TODO: make this all work with sqlalchemy 1.4
       - name: Upgrade sql alchemy
         run: poetry run pip install sqlalchemy==2.0.32
-
-        # run sources tests 
-      - run: poetry run pytest tests/sources 
-        name: Run tests Linux
-        env:
-          DESTINATION__POSTGRES__CREDENTIALS: postgresql://loader:loader@localhost:5432/dlt_data
 
         # run sources tests in load against configured destinations
       - run: poetry run pytest tests/load/sources 
         name: Run tests Linux
         env:
           DESTINATION__POSTGRES__CREDENTIALS: postgresql://loader:loader@localhost:5432/dlt_data
+
