+++ tests/sources/filesystem/test_filesystem_pipeline_template.py
