+++ tests/sources/rest_api/test_rest_api_pipeline_template.py
