+++ tests/sources/sql_database/__init__.py
@@ -1,0 +1,1 @@
+# almost all tests are in tests/load since a postgres instance is required for this to work
